Remove Configurator class methods into dedicated file
The `Configurator` class methods are making the file to long. Instead
they have been moved to a dedicated file that reopens the class and
defines them.
